pyproject.toml
[tool.poetry]
name = "aleph_alpha_showcase_backend"
version = "0.1.0"
description = ""
authors = ["Marc Fabian Mezger <marc.mezger@gmail.com>"]
license = ""
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.1"
uvicorn = "^0.18.3"
aleph-alpha-client = "^2.4.4"
python-dotenv = "^0.21.0"
timm = "^0.6.11"
transformers = "^4.23.1"
opencv-python = "^4.6.0.66"
python-multipart = "^0.0.5"
torch = "1.12.0"
pandas = "^1.5.0"
pre-commit = "^2.20.0"
hydra-core = "^1.2.0"
pytorch-lightning = "^1.7.7"
braceexpand = "^0.1.7"
webdataset = "^0.2.26"
librosa = "^0.9.2"
sentencepiece = "^0.1.97"
cython = "^0.29.32"
openai-whisper = "^20230124"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[settings]
known_third_party = ["PIL", "aleph_alpha_client", "app", "cv2", "fastapi", "pandas", "pydantic", "torch", "transformers", "whisper"]
